+++ ads-in.js
@@ -15,25 +15,30 @@ module.exports = function (RED) {
       node.timezone = config.timezone
       node.inValue = config.inValue||'payload'
       node.useInputMsg = config.useInputMsg||false
+      node.topic = config.topic||''
+      node.hasTopic = node.topic.length > 0
       debug('config:',node)
 
-      node.onAdsData = function (handle){
-        debug('onAdsData:','node.id',node.id,'node.symname',node.symname,'handle.value',handle.value)
-        var msg = {}
-        RED.util.setMessageProperty(msg, node.inValue, handle.value)
-        node.send(msg)
-        debug('onAdsData:','node.id',node.id,'node.symname',node.symname,'msg',msg)
-      }
+      //node.onAdsData = function (handle){
+      //  debug('onAdsData:','node.id',node.id,'node.symname',node.symname,'handle.value',handle.value)
+      //  var msg = {}
+      //  RED.util.setMessageProperty(msg, node.inValue, handle.value)
+      //  node.send(msg)
+      //  debug('onAdsData:','node.id',node.id,'node.symname',node.symname,'msg',msg)
+      //}
 
       this.on("input", function(msg) {
         debug('input:',msg)
         var outMsg = {}
         if (node.useInputMsg) {
-          outMsg = msg
+          outMsg = Object.assign({},msg)
         }
 
         node.adsDatasource.read(node,function (handle){
           RED.util.setMessageProperty(outMsg, node.inValue, handle.value)
+          if (node.hasTopic) {
+            outMsg.topic = node.topic
+          }
           node.send(outMsg)
           debug('input:','node.id',node.id,'node.symname',node.symname,'outMsg',outMsg)
         })