+++ ads-notification.js
@@ -17,12 +17,17 @@ module.exports = function (RED) {
       node.maxDelay = config.maxDelay
       node.cycleTime = config.cycleTime
       node.property = config.property||'payload'
+      node.topic = config.topic||''
+      node.hasTopic = node.topic.length > 0
       debug('config:',node)
 
       node.onAdsData = function (handle){
         debug('onAdsData:','node.id',node.id,'node.symname',node.symname,'handle.value',handle.value)
         var msg = {}
         RED.util.setMessageProperty(msg, node.property, handle.value)
+        if (node.hasTopic) {
+          msg.topic = node.topic
+        }
         node.send(msg)
         debug('onAdsData:','node.id',node.id,'node.symname',node.symname,'msg',msg)
 //        node.setStatus()